pyproject.toml
[project]
name = "OnePaceOrganizer"
version = "1.0.0"
description = "One Pace Organizer for Plex and Jellyfin: Add metadata like summaries, titles and posters to your One Pace collection"
authors = [
    {name = "Isatis",email = "git@isatis.blue"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiopath>=0.7.7",
    "beautifulsoup4>=4.13.4",
    "deepdiff>=8.6.0",
    "enzyme>=0.5.2",
    "httpx-retries>=0.4.0",
    "httpx[http2]>=0.28.1",
    "javaproperties>=0.8.2",
    "loguru>=0.7.3",
    "orjson>=3.11.2",
    "plexapi>=4.17.0",
    "prompt-toolkit>=3.0.52",
    "pyinstaller>=6.15.0",
    "pyside6>=6.9.1",
    "pyyaml>=6.0.2",
    "rss-parser>=2.1.1",
]
